pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "bitvoker"
version = "1.3.0"
description = "bitvoker notification system"
authors = [
    {name = "Rasool Fatemi", email = "rmfatemi@gmail.com"},
]
readme = "README.md"
requires-python = ">=3.11,<4.0"
dependencies = [
    "pyyaml>=6.0",
    "requests>=2.28.0",
    "meta-ai-api>=1.2.5",
    "uvicorn>=0.34.2",
    "fastapi>=0.115.12",
    "apprise>=1.9.3",
]

[project.scripts]
bitvoker = "bitvoker.server:main"

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pre-commit>=4.2.0",
    "deptry>=0.23.0",
]

[tool.black]
target-version = ["py311"]
preview = true
line-length = 120

[tool.ruff]
target-version = "py311"
line-length = 120
include = [
    "bitvoker/**/*.py",
    "tests/**/*.py",
    "*.py",
]
